+++ Semana_3/gobarber-native/gobarberapp/src/pages/SignIn/index.tsx
@@ -1,16 +1,19 @@
-import React from "react";
+import React, { useCallback, useRef } from "react";
 import {
   Image,
   KeyboardAvoidingView,
   Platform,
   View,
   ScrollView,
+  TextInput
 } from "react-native";
 import logoImg from "../../assets/logo.png"; // Essa linha fica com erro ate criar a pasta @types
 import Button from "../../Components/Button";
 import Input from "../../Components/Input";
 import Icon from "react-native-vector-icons/Feather";
-import {useNavigation} from '@react-navigation/native'
+import { useNavigation } from "@react-navigation/native";
+
+import { FormHandles } from "@unform/core";
 import {
   Container,
   Title,
@@ -18,10 +21,23 @@ import {
   ForgotPasswordText,
   CreateAccountButton,
   CreateAccountButtonText,
+  Formulary,
 } from "./styles";
 
+
+
+interface InputPasswordRefProps{
+  value:string
+}
 const SignIn: React.FC = () => {
+
   const navigation = useNavigation();
+  const formRef = useRef<FormHandles>(null);
+  const passwordInputRef = useRef<TextInput>(null);
+  const handleSignIn = useCallback(async (data: object) => {
+    console.log(data);
+  }, []);
+
   return (
     <>
       <KeyboardAvoidingView
@@ -38,15 +54,32 @@ const SignIn: React.FC = () => {
             <View>
               <Title>Faça seu logon</Title>
             </View>
-            <Input name="Email" icon="mail" placeholder="E-mail" />
-            <Input name="Password" icon="lock" placeholder="Senha" />
-            <Button
-              onPress={() => {
-                console.log("Hi");
-              }}
-            >
-              Entrar
-            </Button>
+            <Formulary ref={formRef} onSubmit={handleSignIn}>
+              <Input 
+              returnKeyType="next"
+              onSubmitEditing={() => {passwordInputRef.current?.focus()}}
+              autoCorrect={false}
+              autoCapitalize="none"
+              keyboardType="email-address"
+              name="Email" 
+              icon="mail" 
+              placeholder="E-mail" />
+              <Input
+              ref={passwordInputRef}
+              secureTextEntry
+              returnKeyType="send"
+              onSubmitEditing={() => {formRef.current?.submitForm()}}
+              name="Password"
+              icon="lock"
+              placeholder="Senha" />
+              <Button
+                onPress={() => {
+                  formRef.current?.submitForm();
+                }}
+              >
+                Entrar
+              </Button>
+            </Formulary>
             <ForgotPassword
               onPress={() => {
                 console.log("Password");
@@ -57,7 +90,11 @@ const SignIn: React.FC = () => {
           </Container>
         </ScrollView>
       </KeyboardAvoidingView>
-      <CreateAccountButton onPress={()=>{navigation.navigate('SignUp')}}>
+      <CreateAccountButton
+        onPress={() => {
+          navigation.navigate("SignUp");
+        }}
+      >
         <>
           <Icon name="log-in" size={20} color="#FF9000" />
           <CreateAccountButtonText>Criar uma conta</CreateAccountButtonText>