+++ Semana_3/gobarber-native/gobarberapp/src/pages/SignIn/styles.ts
@@ -2,6 +2,7 @@ import styled from "styled-components/native";
 import { Platform } from "react-native";
 import { fontRegular } from "../../assets/fonts";
 import { getBottomSpace } from "react-native-iphone-x-helper";
+import { Form } from "@unform/mobile";
 
 export const Container = styled.View`
   flex: 1;
@@ -45,3 +46,7 @@ export const CreateAccountButtonText = styled.Text`
   font-family: ${fontRegular};
   margin-left: 16px;
 `;
+
+export const Formulary = styled(Form)`
+    width: 100%;
+`