+++ Semana_3/gobarber-native/gobarberapp/src/pages/SignUp/index.tsx
@@ -1,20 +1,30 @@
-import React from "react";
+import React, { useCallback, useRef } from "react";
 import {
   Image,
   KeyboardAvoidingView,
   Platform,
   View,
   ScrollView,
+  TextInput,
 } from "react-native";
 import logoImg from "../../assets/logo.png"; // Essa linha fica com erro ate criar a pasta @types
 import Button from "../../Components/Button";
 import Input from "../../Components/Input";
 import Icon from "react-native-vector-icons/Feather";
 import { useNavigation } from "@react-navigation/native";
+import { Form } from "@unform/mobile";
+import { FormHandles } from "@unform/core";
 import { Container, Title, BackToSignIn, BackToSignInText } from "./styles";
 
 const SignUp: React.FC = () => {
   const navigation = useNavigation();
+  const formRef = useRef<FormHandles>(null);
+  let emailInputRef = useRef<TextInput>(null);
+  let passwordInputRef = useRef<TextInput>(null);
+  const handleSubmit = useCallback((data: object) => {
+    console.log(data);
+  }, []);
+
   return (
     <>
       <KeyboardAvoidingView
@@ -31,12 +41,48 @@ const SignUp: React.FC = () => {
             <View>
               <Title>Crie sua conta</Title>
             </View>
-            <Input name="name" icon="user" placeholder="Nome" />
-            <Input name="Email" icon="mail" placeholder="E-mail" />
-            <Input name="Password" icon="lock" placeholder="Senha" />
+            <Form ref={formRef} onSubmit={handleSubmit}>
+              <Input
+                autoCorrect={true}
+                autoCapitalize="words"
+                name="name"
+                icon="user"
+                placeholder="Nome"
+                enablesReturnKeyAutomatically={true}
+                returnKeyType="next"
+                onSubmitEditing={() => {
+                  emailInputRef.current?.focus();
+                }}
+              />
+              <Input
+                ref={emailInputRef}
+                keyboardType="email-address"
+                autoCorrect={false}
+                autoCapitalize="none"
+                name="Email"
+                icon="mail"
+                placeholder="E-mail"
+                enablesReturnKeyAutomatically={true}
+                returnKeyType="next"
+                onSubmitEditing={() => {
+                  passwordInputRef.current?.focus();
+                }}
+              />
+              <Input
+                ref={passwordInputRef}
+                name="Password"
+                icon="lock"
+                placeholder="Senha"
+                secureTextEntry
+                textContentType={"newPassword"}
+                returnKeyType="send"
+                enablesReturnKeyAutomatically={true}
+                onSubmitEditing={() => formRef.current?.submitForm()}
+              />
+            </Form>
             <Button
               onPress={() => {
-                console.log("Hi");
+                formRef.current?.submitForm();
               }}
             >
               Entrar
@@ -44,7 +90,7 @@ const SignUp: React.FC = () => {
           </Container>
         </ScrollView>
       </KeyboardAvoidingView>
-      <BackToSignIn onPress={()=> navigation.goBack()}>
+      <BackToSignIn onPress={() => navigation.goBack()}>
         <>
           <Icon name="arrow-left" size={20} color="#FFF" />
           <BackToSignInText>Voltar para Logon</BackToSignInText>